Hotfix: epicure dashboard repository and service

+++ src/main/java/acme/entities/epicureDashboard/EpicureDashboard.java
@@ -17,8 +17,8 @@ public class EpicureDashboard implements Serializable{
 	protected Map<StatusType,Integer> totalFineDish;
 	protected Map<StatusType,Double> averageFineDishRetailPrice;
 	protected Map<StatusType,Double> deviationFineDishRetailPrice;
-	protected Map<StatusType,Integer> maximumFineDishRetailPrice;
-	protected Map<StatusType,Integer> minimumUtensilRetailPrice;
+	protected Map<StatusType,Double> maximumFineDishRetailPrice;
+	protected Map<StatusType,Double> minimumUtensilRetailPrice;
